Stochastic Optimisation Test Bed
This is a “black-box” simulation environment that is constructed around
three main modules of: System, Disturbance and Control. The intention
is to fix two of the modules (eg. System and Disturbance) and be able
to bench-mark the performance of a range of different modules for the
remaining component (eg. compare performance of different Control
Algorithms).